Refactor: Extract DiscoveryCandidates.ts into focused modules (#1473)
Decompose the 2,227-line DiscoveryCandidates.ts into 5 focused,
single-responsibility modules, reducing the coordinator to 387 lines
(well under the 600-line target).

Extracted modules:
- CombinedCandidates.ts (845 lines) — multi-step combination strategies
- CandidateApplication.ts (615 lines) — validation and creature modification
- CandidateCreation.ts (483 lines) — single-step candidate builders
- CandidateDescriptions.ts (155 lines) — emoji and description generation
- CandidateScoring.ts (77 lines) — expected improvement calculations

Bonus: replaced `any` types with `CreatureExport` in CandidateApplication.ts
for improved type safety.

All 3,823 tests pass. No functional changes.
